Moved test directory and updated Neural CDE.
- Tests have been moved into the package directory due to problems
with root and package being named the same.
- Neural CDE now performs the interpolation inside the class, so we hand
it normal format data. This should be ok due to us sticking to linear so
the time cost is small, but it saves on data processing annoyances.
- Neural CDE does not need to be passed an initial value any longer,
this is all done within the forward call, the only additional data is
static data which is specified like before.